streamlit/requirements.txt
streamlit==1.38.0
requests==2.32.3
pandas==2.2.2
python-slugify
